Added the path selection popup
