+++ WheelWizard/Views/Popups/Generic/FirstTimeSetupPopup.axaml.cs
@@ -1,14 +1,5 @@
-using System.Runtime.InteropServices;
-using Avalonia;
-using Avalonia.Controls;
 using Avalonia.Interactivity;
-using Avalonia.Layout;
-using Avalonia.Media;
-using Avalonia.Platform.Storage;
 using Microsoft.Extensions.Logging;
-using WheelWizard.DolphinManagement.Abstractions;
-using WheelWizard.Helpers;
-using WheelWizard.Services;
 using WheelWizard.Settings;
 using WheelWizard.Shared.DependencyInjection;
 using WheelWizard.Views.Popups.Base;
@@ -17,33 +8,33 @@ namespace WheelWizard.Views.Popups.Generic;
 
 public partial class FirstTimeSetupPopup : PopupContent
 {
-    private sealed record DolphinCandidate(string DisplayName, string? Path, bool Found);
+    private enum SetupStep
+    {
+        Dolphin,
+        GamePaths,
+    }
 
     private readonly TaskCompletionSource<bool> _completionSource = new();
     private bool _setupCompleted;
 
-    private string? _selectedDolphinTarget;
+    private SetupStep _currentStep = SetupStep.Dolphin;
 
     [Inject]
     private ILogger<FirstTimeSetupPopup> Logger { get; set; } = null!;
 
     [Inject]
-    private IDolphinLocator DolphinLocator { get; set; } = null!;
-
-    [Inject]
-    private IDolphinInstaller DolphinInstaller { get; set; } = null!;
-
-    [Inject]
     private ISettingsManager SettingsService { get; set; } = null!;
+
+    public bool WasSkipped { get; private set; }
 
     public FirstTimeSetupPopup()
         : base(true, false, false, "Wheel Wizard")
     {
         InitializeComponent();
 
-        PlatformTextBlock.Text = $"Detected platform: {GetPlatformName()}";
-        //BrowseButton.IsEnabled = EnvHelper.IsFlatpakSandboxed(); Not sure how to check for flatpak installation because this doesnt really work
-        PopulateDetectedLocations();
+        InitializeDolphinStep();
+        ShowStep(SetupStep.Dolphin);
+        AutoDetectPaths();
     }
 
     public Task<bool> ShowAndAwaitCompletionAsync()
@@ -52,177 +43,65 @@ public partial class FirstTimeSetupPopup : PopupContent
         return _completionSource.Task;
     }
 
-    private static string GetPlatformName()
+    private void ShowStep(SetupStep step)
     {
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            return "Windows";
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            return "macOS";
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            return "Linux";
-        return "Unknown";
+        _currentStep = step;
+
+        DolphinStepPanel.IsVisible = step == SetupStep.Dolphin;
+        GamePathsStepPanel.IsVisible = step == SetupStep.GamePaths;
+
+        BackButton.IsVisible = step == SetupStep.GamePaths;
+        SkipButton.IsVisible = step == SetupStep.GamePaths;
+
+        UpdateContinueState();
     }
 
-    private void PopulateDetectedLocations()
+    private void UpdateContinueState()
     {
-        DetectedLocationsPanel.Children.Clear();
-
-        var candidates = DolphinLocator.DetectInstallations();
-        if (candidates.Count == 0)
+        ContinueButton.IsEnabled = _currentStep switch
         {
-            DetectedLocationsPanel.Children.Add(
-                new TextBlock
-                {
-                    Classes = { "BodyText" },
-                    Opacity = 0.75,
-                    TextWrapping = TextWrapping.Wrap,
-                    Text = "No Dolphin installations were detected automatically. Please select one manually below.",
-                }
-            );
-            return;
-        }
-
-        foreach (var candidate in candidates)
-            DetectedLocationsPanel.Children.Add(CreateCandidateRow(candidate));
-    }
-
-    private Grid CreateCandidateRow(DolphinInstallation candidate)
-    {
-        var subtitle = candidate.Found ? candidate.LaunchTarget ?? string.Empty : "Not found on this system";
-
-        var content = new StackPanel { Spacing = 2 };
-        content.Children.Add(
-            new TextBlock
-            {
-                Classes = { "BodyText" },
-                FontWeight = FontWeight.SemiBold,
-                Text = candidate.DisplayName,
-            }
-        );
-        content.Children.Add(
-            new TextBlock
-            {
-                Classes = { "BodyText" },
-                Opacity = 0.7,
-                TextWrapping = TextWrapping.Wrap,
-                Text = subtitle,
-            }
-        );
-
-        var radio = new RadioButton
-        {
-            GroupName = "DolphinLocation",
-            Content = content,
-            IsEnabled = candidate.Found,
-            Tag = candidate.LaunchTarget,
-            HorizontalContentAlignment = HorizontalAlignment.Stretch,
+            SetupStep.Dolphin => IsDolphinStepComplete,
+            SetupStep.GamePaths => IsGamePathsStepComplete,
+            _ => false,
         };
-        radio.IsCheckedChanged += DetectedLocation_OnChecked;
-
-        var row = new Grid { ColumnDefinitions = new ColumnDefinitions("*,Auto") };
-        row.Children.Add(radio);
-
-        if (!candidate.Found && candidate.DisplayName.Contains("flatpak", StringComparison.OrdinalIgnoreCase))
-        {
-            var installButton = new Components.Button
-            {
-                Variant = Components.Button.ButtonsVariantType.Default,
-                Text = "Install",
-                VerticalAlignment = VerticalAlignment.Center,
-                Margin = new Thickness(8, 0, 0, 0),
-            };
-            installButton.Click += InstallFlatpakButton_OnClick;
-            Grid.SetColumn(installButton, 1);
-            row.Children.Add(installButton);
-        }
-
-        return row;
-    }
-
-    private async void InstallFlatpakButton_OnClick(object? sender, RoutedEventArgs e)
-    {
-        var progressWindow = new ProgressWindow()
-            .SetGoal(t("progress.installing_dolphin"))
-            .SetExtraText(t("progress.this_may_take_a_while"));
-        progressWindow.Show();
-        var progress = new Progress<int>(progressWindow.UpdateProgress);
-        var installResult = await DolphinInstaller.InstallDolphin(progress);
-        progressWindow.Close();
-        if (installResult.IsFailure)
-        {
-            await new MessageBoxWindow()
-                .SetMessageType(MessageBoxWindow.MessageType.Error)
-                .SetTitleText("Failed to install Dolphin")
-                .SetInfoText(installResult.Error.Message)
-                .ShowDialog();
-            return;
-        }
-
-        // Reload all radio buttons
-        var candidates = DolphinLocator.DetectInstallations();
-        foreach (var radio in DetectedLocationsPanel.Children.SelectMany(GetRadioButtons))
-        {
-            if (radio?.Tag is null)
-                continue;
-            var match = candidates.FirstOrDefault(c => radio.Tag.Equals(c.LaunchTarget));
-            if (match != null)
-            {
-                radio.IsEnabled = match.Found;
-            }
-        }
-    }
-
-    private static IEnumerable<RadioButton> GetRadioButtons(Control row) => row is Panel panel ? panel.Children.OfType<RadioButton>() : [];
-
-    private void DetectedLocation_OnChecked(object? sender, RoutedEventArgs e)
-    {
-        if (sender is not RadioButton { IsChecked: true } radio)
-            return;
-
-        ManualPathTextBox.Text = string.Empty;
-        SetSelectedTarget(radio.Tag as string);
-    }
-
-    private async void BrowseButton_OnClick(object? sender, RoutedEventArgs e)
-    {
-        try
-        {
-            var path = await OpenDolphinFilePickerAsync(this);
-            Logger.LogInformation(path);
-            if (string.IsNullOrWhiteSpace(path))
-                return;
-
-            // A manual pick wins: clear any detected radio selection.
-            foreach (var radio in DetectedLocationsPanel.Children.SelectMany(GetRadioButtons))
-                radio.IsChecked = false;
-
-            ManualPathTextBox.Text = path;
-            SetSelectedTarget(path);
-        }
-        catch (Exception ex)
-        {
-            Logger.LogError(ex, "Failed to pick a Dolphin location.");
-            ShowError("Something went wrong while selecting the file.");
-        }
-    }
-
-    private void SetSelectedTarget(string? target)
-    {
-        _selectedDolphinTarget = string.IsNullOrWhiteSpace(target) ? null : target;
-        ContinueButton.IsEnabled = _selectedDolphinTarget != null;
         ErrorTextBlock.IsVisible = false;
     }
 
     private void ContinueButton_OnClick(object? sender, RoutedEventArgs e)
     {
-        if (string.IsNullOrWhiteSpace(_selectedDolphinTarget))
+        if (_currentStep == SetupStep.Dolphin)
         {
-            ShowError("Please select a Dolphin installation to continue.");
+            if (!IsDolphinStepComplete)
+            {
+                ShowError("Please select a Dolphin installation to continue.");
+                return;
+            }
+
+            SaveDolphinStep();
+            ShowStep(SetupStep.GamePaths);
             return;
         }
 
-        SettingsService.Set(SettingsService.DOLPHIN_LOCATION, _selectedDolphinTarget);
+        if (!IsGamePathsStepComplete)
+        {
+            ShowError("Please select both the Dolphin user folder and your Mario Kart Wii game file.");
+            return;
+        }
 
+        SaveGamePathsStep();
+        CompleteSetup();
+    }
+
+    private void BackButton_OnClick(object? sender, RoutedEventArgs e) => ShowStep(SetupStep.Dolphin);
+
+    private void SkipButton_OnClick(object? sender, RoutedEventArgs e)
+    {
+        WasSkipped = true;
+        CompleteSetup();
+    }
+
+    private void CompleteSetup()
+    {
         _setupCompleted = true;
         _completionSource.TrySetResult(true);
         Close();
@@ -236,21 +115,5 @@ public partial class FirstTimeSetupPopup : PopupContent
     {
         ErrorTextBlock.Text = message;
         ErrorTextBlock.IsVisible = true;
-    }
-
-    private static async Task<string?> OpenDolphinFilePickerAsync(Visual owner)
-    {
-        var executableFileType = new FilePickerFileType("Executable files")
-        {
-            Patterns = Environment.OSVersion.Platform switch
-            {
-                PlatformID.Win32NT => new[] { "*.exe" },
-                PlatformID.Unix => new[] { "*", "*.sh" },
-                PlatformID.MacOSX => new[] { "*", "*.app" },
-                _ => new[] { "*" }, // Fallback
-            },
-        };
-        var filePath = await FilePickerHelper.OpenSingleFileAsync("Select the Dolphin executable", [executableFileType], owner);
-        return filePath;
     }
 }
